Server/pyproject.toml
[project]
name = "mac"
version = "0.1.0"
description = "E-commerce Backend with Admin Dashboard using Supabase (PostgreSQL)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "pillow>=10.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.1.0",
    "razorpay>=1.4.1",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "supabase>=2.3.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]
